requirements.txt
anthropic>=0.42.0
streamlit>=1.40.0
Pillow>=10.0.0
